0510_완료

+++ src/main/java/com/estsoft/jblog/service/CategoryService.java
@@ -18,7 +18,7 @@ public class CategoryService {
 		categoryDao.create(blog_id);
 	}
 	
-	public int insertCategory(CategoryVO vo) {	
+	public Long insertCategory(CategoryVO vo) {	
 		return categoryDao.insertCategory(vo);
 	}
